Split the Fidelity converter from its React options component
The import extension reuses the client's CSV converters so parsing logic is not
duplicated, but it has no UI framework. fidelity.tsx mixed a pure converter with
a React <select>, and the module-scope register() call referenced the component,
so importing the converter pulled React into the bundle.

The converter moves to fidelity-csv.ts, named for the format id it implements.
fidelity.tsx keeps the options component and the registry entry, so brokers still
self-register on import of the barrel and the upload modal is unchanged. The
other converters needed no change: standard.ts and ibkr-ofx.ts contain no React,
and registry.ts references it only through an import type that is erased.

The two modules deliberately do not share a basename. Next resolves .tsx before
.ts and Vite resolves .ts before .tsx, so a fidelity.ts alongside a fidelity.tsx
would have registered under Next and silently not registered under Vitest.

Also exports FIDELITY_TYPE_TO_OFX, so a caller that uploads despite dropped rows
can name the transaction types it dropped, and adds an optional headers argument
to the gRPC-Web transport. The browser app authenticates by cookie and needs no
headers; a client whose request is cross-site must send the session id as
Authorization: Bearer, which the transport previously had no way to express.
